app/requirements.txt
# pymongo 
# pymongo[srv] 
numpy
openai
openpyxl
langchain-community
langchain-huggingface
unstructured[all-docs]
langchain-chroma
transformers
torch
scikit-learn
fastapi 
uvicorn
gunicorn
pandas
streamlit
pymilvus

PyMuPDF